pyproject.toml
[tool.poetry]
name = "datadoc-example"
version = "0.1.0"
description = ""
authors = ["trygu <trygve.falch@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
dapla-toolbelt = ">=1.3.2"
ipykernel = ">=6.15.3"
ssb-datadoc-model = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.1.3"
pytest-cov = ">=5.0.0"
black = {extras = ["jupyter"], version = ">=24.4.2"}
isort = ">=5.12.0"
pre-commit = ">=3.3.1"
pre-commit-hooks = ">=4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
force_single_line = true
skip_gitignore = true
lines_after_imports = 2
# The line below is needed for jupyter notebooks stored as .py in percent format.
# See https://github.com/PyCQA/isort/issues/1338 for details
treat_comments_as_code = ["# %%"]

[tool.jupytext]
formats = "ipynb,auto:percent"
notebook_metadata_filter = "jupytext.text_representation,-jupytext.text_representation.jupytext_version,-widgets,-varInspector"
cell_metadata_filter = "-papermill,tags"

[tool.pytest.ini_options]
pythonpath = ["src"]
